Added controller for login
