+++ routes/Route.js
@@ -18,76 +18,7 @@ router.get('/', (req, res)=>{
 router.post('/userExists', userController.userExists);  //username exists or not
 router.get('/verify/:authToken',userController.verifyToken);  // verify the auth token
 router.post('/signup', userController.signup);
-
-//done
-router.post('/login', (req, res) => {
-
-    var username = req.body.username;
-    var password = req.body.password;
-
-    User.find({ username: username })
-        .exec()
-        .then(result => {
-            if (result.length) {
-                var email = result[0].email;
-
-                console.log(result);
-                console.log(password);
-                console.log(result[0].password);
-                console.log(email);
-
-                //password match and jwt generation
-                bcrypt.compare(password, result[0].password)
-                    .then(comparedResult => {
-                        const uniqueString = jwt.sign({ email, username }, process.env.JWT_ACC_ACTIVATE1, { expiresIn: '2h' });
-                        console.log(uniqueString);
-                        console.log(comparedResult);
-
-                        if (comparedResult) {
-                            if (!result[0].isVerified) {
-                                res.status(200).json({
-                                    message: "Please vefify your account to login",
-                                    result: "false"
-                                })
-                            }
-                            else {
-                                res.status(200).json({
-                                    message: uniqueString,
-                                    result: "true"
-                                })
-                            }
-                        }
-                        else {
-                            res.status(200).json({
-                                message: "Incorrect username or password",
-                                result: "false"
-                            })
-                        }
-                    })
-                    .catch(error => {
-                        console.log(error);
-                        res.status(500).json({
-                            message: error,
-                        })
-                    })
-
-
-            }
-            else {
-                console.log(username);
-                res.status(200).json({
-                    message: "Incorrect username or password",
-                    result: "false"
-                })
-            }
-        })
-        .catch(error => {
-            console.log(error);
-            res.status(500).json({
-                message: error
-            });
-        });
-});
+router.post('/login', userController.login);
 
 //done 
 router.post('/displayLogs', (req, res) => {
